+++ src/main/java/care/smith/top/top_phenotypic_query/c2reasoner/constants/E.java
@@ -1,7 +1,5 @@
 package care.smith.top.top_phenotypic_query.c2reasoner.constants;
 
-import java.math.BigDecimal;
-
 import care.smith.top.model.Constant;
 import care.smith.top.model.DataType;
 import care.smith.top.model.NumberConstant;
@@ -23,7 +21,7 @@ public class E extends ConstantEntity {
   private static Value createValue() {
     return new NumberValue()
         .value(
-            new BigDecimal(
+            new Double(
                 "2.71828182845904523536028747135266249775724709369995957496696762772407663"))
         .dataType(DataType.NUMBER);
   }
